Delivery : * Merge files * delete already merged file
